requirements.txt
pika
flask
mysql.connector
boto3
mysql-connector-python
requests
